Added Requests
Se añaden requests GET, POST y DELETE para los empleados, clientes y citas.
Falta update para editar los objetos existentes.

+++ API/Models/Client.cs
@@ -2,7 +2,7 @@ namespace API.Models;
 
 public class Client
 {
-    public long idNumber{get; set;}
+    public int idNumber{get; set;}
     public string name{get; set;}
     public string lastName{get; set;}
     public string secondLastName{get; set;}
